+++ cloudformation/emrserverless/aws-emrserverless-application_initialcapacityconfig.go
@@ -18,7 +18,7 @@ type Application_InitialCapacityConfig struct {
 	// WorkerCount AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-emrserverless-application-initialcapacityconfig.html#cfn-emrserverless-application-initialcapacityconfig-workercount
-	WorkerCount int `json:"WorkerCount"`
+	WorkerCount int64 `json:"WorkerCount"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
