+++ cloudformation/iotwireless/aws-iotwireless-taskdefinition_lorawanupdategatewaytaskcreate.go
@@ -18,7 +18,7 @@ type TaskDefinition_LoRaWANUpdateGatewayTaskCreate struct {
 	// SigKeyCrc AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-iotwireless-taskdefinition-lorawanupdategatewaytaskcreate.html#cfn-iotwireless-taskdefinition-lorawanupdategatewaytaskcreate-sigkeycrc
-	SigKeyCrc *int `json:"SigKeyCrc,omitempty"`
+	SigKeyCrc *int64 `json:"SigKeyCrc,omitempty"`
 
 	// UpdateSignature AWS CloudFormation Property
 	// Required: false
