+++ cloudformation/kafkaconnect/aws-kafkaconnect-connector_customplugin.go
@@ -18,7 +18,7 @@ type Connector_CustomPlugin struct {
 	// Revision AWS CloudFormation Property
 	// Required: true
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-kafkaconnect-connector-customplugin.html#cfn-kafkaconnect-connector-customplugin-revision
-	Revision int `json:"Revision"`
+	Revision int64 `json:"Revision"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
