+++ cloudformation/lightsail/aws-lightsail-distribution_cachesettings.go
@@ -23,7 +23,7 @@ type Distribution_CacheSettings struct {
 	// DefaultTTL AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-lightsail-distribution-cachesettings.html#cfn-lightsail-distribution-cachesettings-defaultttl
-	DefaultTTL *int `json:"DefaultTTL,omitempty"`
+	DefaultTTL *int64 `json:"DefaultTTL,omitempty"`
 
 	// ForwardedCookies AWS CloudFormation Property
 	// Required: false
@@ -43,12 +43,12 @@ type Distribution_CacheSettings struct {
 	// MaximumTTL AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-lightsail-distribution-cachesettings.html#cfn-lightsail-distribution-cachesettings-maximumttl
-	MaximumTTL *int `json:"MaximumTTL,omitempty"`
+	MaximumTTL *int64 `json:"MaximumTTL,omitempty"`
 
 	// MinimumTTL AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-lightsail-distribution-cachesettings.html#cfn-lightsail-distribution-cachesettings-minimumttl
-	MinimumTTL *int `json:"MinimumTTL,omitempty"`
+	MinimumTTL *int64 `json:"MinimumTTL,omitempty"`
 
 	// AWSCloudFormationDeletionPolicy represents a CloudFormation DeletionPolicy
 	AWSCloudFormationDeletionPolicy policies.DeletionPolicy `json:"-"`
